Absorb voltdbipc.h into voltdbipc.cpp for ease of maintenance.
